Move goal vision read models into control plane
